% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie_AIC.R
\name{tweedie_AIC}
\alias{tweedie_AIC}
\alias{AICtweedie}
\title{AIC for Tweedie Glms}
\usage{
tweedie_AIC(glm.obj, dispersion = NULL, k = 2, verbose = TRUE)

AICtweedie(glm.obj, dispersion = NULL, k = 2, verbose = TRUE)
}
\arguments{
\item{glm.obj}{a fitted \code{glm} object, fitted using the \code{tweedie} family.}

\item{dispersion}{the dispersion parameter, usually extracted from \code{glm.obj}; however, occasionally  a specified value of the dispersion may be needed.}

\item{k}{the \acronym{aic} penalty; \code{k = 2} (the default) produces the AIC.}

\item{verbose}{logical; if \code{TRUE}, display details of the internal process. The default is \code{FALSE}.}
}
\value{
The value of the computed \acronym{aic}.
}
\description{
Evaluates the \acronym{aic} for a fitted Tweedie \acronym{glm}.
The Tweedie family of distributions belong to the class of exponential dispersion models (\acronym{edm}s), 
famous for their role in generalized linear models. 
The Tweedie distributions are the \acronym{edm}s with a variance of the form 
\eqn{\mbox{var}[Y] = \phi\mu^p}{var[Y] = phi*mu^p} where \eqn{p \ge 1}{p >= 1}.
\emph{This function only evaluates for \eqn{p \ge 1}{p >= 1}.}
}
\details{
The \acronym{aic} is computed by evaluating the density function.
}
\note{
Evaluating the likelihood can be time consuming, so the function may take some time for large data sets.
}
\examples{
# Fit a Tweedie density using  tweedie  family function from  statmod
pretend <- data.frame( y = stats::rgamma(20, shape = 1, rate = 1) )
fit <- glm(y ~ 1, data = pretend, 
           family = statmod::tweedie(link.power = 0, var.power = 2.1))

# Compute the AIC
tweedie_AIC(fit)

}
\references{
Dunn, P. K. and Smyth, G. K. (2008).
Evaluation of Tweedie exponential dispersion model densities by Fourier inversion.
\emph{Statistics and Computing}, 
\bold{18}, 73--86.
\doi{10.1007/s11222-007-9039-6}

Dunn, Peter K and Smyth, Gordon K (2005).
Series evaluation of Tweedie exponential dispersion model densities
\emph{Statistics and Computing},
\bold{15}(4). 267--280.
\doi{10.1007/s11222-005-4070-y}

Jorgensen, B. (1997).
\emph{Theory of Dispersion Models}.
Chapman and Hall, London.

Sakamoto, Y., Ishiguro, M., and Kitagawa G. (1986). 
\emph{Akaike Information Criterion Statistics}. 
D. Reidel Publishing Company.
}
\seealso{
\code{\link{dtweedie}}
}
