% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie-package.R
\docType{package}
\name{tweedie-package}
\alias{tweedie-package}
\alias{tweedie}
\title{Evaluation of Tweedie Exponential Family Models}
\description{
This package provides maximum likelihood computations for Tweedie 
    families, including the series expansion (Dunn and Smyth, 2005) and the 
    Fourier inversion (Dunn and Smyth, 2008), and related methods.
}
\author{
\strong{Maintainer}: Peter K. Dunn \email{pdunn2@usc.edu.au}

}
\keyword{distribution}
\keyword{models}
\keyword{regression}
