% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedie_Plot.R
\name{tweedie_Plot}
\alias{tweedie_Plot}
\title{Tweedie Distributions: Plot Tweedie density functions}
\usage{
tweedie_Plot(
  y,
  xi = NULL,
  mu,
  phi,
  type = "pdf",
  power = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{y}{The values for \eqn{y}{y} in the plot.}

\item{xi}{A synonym for \code{power}.}

\item{mu}{The mean of the distribution.}

\item{phi}{The dispersion parameter.}

\item{type}{The type of plot, either \code{PDF} (the default) or \code{CDF}.}

\item{power}{The variance power.}

\item{add}{If \code{TRUE}, the plot is added to the current plot; if \code{FALSE} (the default) the plot is produced on a fresh plot.}

\item{...}{Plotting parameters}
}
\description{
This function produced a plot of the specified Tweedie distribution.
}
\details{
Plot the Tweedie PDF.

If \eqn{1 < p < 2}{1 < power < 2}, the mass at \eqn{Y=0}{Y = 0} is automatically added.
}
